requirements.txt
python-telegram-bot==13.15
supabase==1.0.3
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0 